move translations to a different file
